Change submodules structure: array, random
