+++ lib/array.js
@@ -1,10 +1,5 @@
 'use strict';
 
-const random = (min, max) => {
-  const [a, b] = max === undefined ? [0, min] : [min, max];
-  return a + Math.floor(Math.random() * (b - a + 1));
-};
-
 const sample = (array, random = Math.random) => {
   const index = Math.floor(random() * array.length);
   return array[index];
@@ -20,4 +15,4 @@ const shuffle = (array, random = Math.random) => {
   return array;
 };
 
-module.exports = { random, sample, shuffle };
+module.exports = { sample, shuffle };
