+++ lib/crypto.js
@@ -28,6 +28,11 @@ const cryptoRandom = () => {
   return buf.readUInt32LE(0) / (UINT32_MAX + 1);
 };
 
+const random = (min, max) => {
+  const [a, b] = max === undefined ? [0, min] : [min, max];
+  return a + Math.floor(Math.random() * (b - a + 1));
+};
+
 const generateUUID = crypto.randomUUID;
 
 const generateKey = (length, possible) => {
@@ -153,6 +158,7 @@ const getX509names = (cert) => {
 
 module.exports = {
   cryptoRandom,
+  random,
   generateUUID,
   generateKey,
   crcToken,
