+++ metautil.js
@@ -1,6 +1,7 @@
 'use strict';
 
 module.exports = {
+  ...require('./lib/array.js'),
   ...require('./lib/async.js'),
   ...require('./lib/crypto.js'),
   ...require('./lib/datetime.js'),
@@ -10,7 +11,6 @@ module.exports = {
   ...require('./lib/network.js'),
   ...require('./lib/objects.js'),
   ...require('./lib/pool.js'),
-  ...require('./lib/random.js'),
   ...require('./lib/semaphore.js'),
   ...require('./lib/strings.js'),
   ...require('./lib/units.js'),
