+++ test/crypro.js
@@ -62,6 +62,15 @@ metatests.case(
       ['secret', 'WRONG-STRING', false],
       ['secret', '', false],
     ],
+    'metautil.random': [
+      [0, 10, (result) => result >= 0 && result <= 10],
+      [1, 10, (result) => result >= 1 && result <= 10],
+      [-1, 10, (result) => result >= -1 && result <= 10],
+      [10, 20, (result) => result >= 10 && result <= 20],
+      [10, 0, (result) => result >= 0 && result <= 10],
+      [20, (result) => result >= 0 && result <= 20],
+      [10, 10, 10],
+    ],
   },
 );
